MNT Begin moving shared eLog code to separate module
